+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changed to use the [Bitrise v0.1 API](https://devcenter.bitrise.io/en/api.html),
 
 - Support endpoint for triggering a build
 - Support aborting a build
+- Support listing registered test devices of all members of a specified Bitrise app
 
 ## v0.1.0